performance progress

+++ Models/MonthlyRecord.cs
@@ -7,11 +7,11 @@ namespace EmployeePerformanceSystem.Models
     {
         public int id { get; set; }
         public int user_id { get; set; }
-        public int month { get; set; }
-        public int? work { get; set; }
-        public int? vacation { get; set; }
-        public int? vacation_sick { get; set; }
-        public int? mission { get; set; }
+        public byte month { get; set; } // tinyint در SQL Server معادل byte در C#
+        public byte? work { get; set; }
+        public byte? vacation { get; set; }
+        public byte? vacation_sick { get; set; }
+        public byte? mission { get; set; }
         public int? overtime_system { get; set; }
         public int? overtime_final { get; set; }
         public int sum_work { get; set; }
